client-server/requirements_ev3.txt
python-ev3dev2>=2.1.0
opencv-python>=4.5.0
numpy>=1.19.0
